requirements.txt
flake8
django-extensions
Werkzeug
django-debug-toolbar
